refactor: service products

+++ database/migrations/2026_08_10_022851_create_service_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('service_id')->constrained()->cascadeOnDelete();
             $table->foreignId('product_id')->constrained()->cascadeOnDelete();
-            $table->integer('quantity')->default(1);
+            $table->decimal('quantity')->default(1);
             $table->timestamps();
         });
     }
